+++ cookbooks/chub_census/roles/census.rb
@@ -1,2 +1,2 @@
 name "census"
-run_list "recipe[chub_sensu_sso::client]", "recipe[chub_census::default]"
+run_list "recipe[chub_census::default]"