+++ environments/ssoprod1.rb
@@ -3,11 +3,11 @@ description "SSO Prod environment 1"
 cookbook_versions({
   "apache2" => "= 1.7.0",
   "apt" => "= 2.3.8",
-  "base" => "= 0.1.12",
+  "base" => "= 0.1.15",
   "bluepill" => "= 2.3.1",
   "build-essential" => "= 1.4.2",
   "chef-client" => "= 3.0.6",
-  "chef-kick" => "= 0.1.0",
+  "chef-kick" => "= 0.1.1",
   "chef_handler" => "= 1.1.5",
   "chub_castle" => "= 0.1.17",
   "chub_census" => "= 0.1.4",